Cleaning up branch
